adjust new chat button display in chat folder
* Display it at the top instead at the end.
This because chats are order by recently created.
UX issue when clicking to the button at the end,
and have to scroll at the top to see and edit newly created chat.

* Display the button when hovering the folder,
not just the group of chats in the folder.
This make the button to display when hover the folder,
even if there is not yet chats list in the folder.